# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password (step 1 of 2FA)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Session (weak when TOTP is enrolled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "400":
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/user:
    post:
      operationId: getUser
      tags: [auth]
      summary: Load the current user for a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/totp/verify:
    post:
      operationId: verifyTotp
      tags: [auth]
      summary: Complete 2FA with a TOTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpVerifyInput"
      responses:
        "200":
          description: Verified session tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpVerifyResponse"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpVerifyResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: Sign-out result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /rag/search:
    post:
      operationId: searchRag
      tags: [rag]
      summary: Single-shot compliance search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RagSearchInput"
      responses:
        "200":
          description: RAG answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagAnswer"
        "400":
          description: Empty query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rag/chat:
    post:
      operationId: chatRag
      tags: [rag]
      summary: RAG assistant chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RagChatInput"
      responses:
        "200":
          description: RAG answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagAnswer"
        "400":
          description: Empty question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rag/storage/sync:
    post:
      operationId: syncRagStorage
      tags: [rag]
      summary: Run the Storage-to-pgvector sweep on demand
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageSyncResult"
        "500":
          description: Sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        category:
          type: string
          enum: [CONFIGURATION, JIRA, SUPABASE, UNKNOWN]
...
    AuthUser:
      type: object
      required: [id]
      properties:
        id: { type: string }
        email: { type: string }
        app_metadata: { type: object, additionalProperties: true }
        user_metadata: { type: object, additionalProperties: true }
    AuthFactor:
      type: object
      required: [id, factor_type, status]
      properties:
        id: { type: string }
        factor_type: { type: string }
        status: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    LoginResult:
      type: object
      required: [ok, weak_session, factors]
      properties:
        ok: { type: boolean }
        access_token: { type: string }
        refresh_token: { type: string }
        user:
          $ref: "#/components/schemas/AuthUser"
        weak_session: { type: boolean }
        factors:
          type: array
          items:
            $ref: "#/components/schemas/AuthFactor"
        message: { type: string }
    UserInput:
      type: object
      required: [access_token]
      properties:
        access_token: { type: string }
    UserResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        user:
          $ref: "#/components/schemas/AuthUser"
        message: { type: string }
    TotpVerifyInput:
      type: object
      required: [factor_id, code, access_token]
      properties:
        factor_id: { type: string }
        code: { type: string }
        access_token: { type: string }
    TotpVerifyResponse:
      type: object
      required: [ok, message]
      properties:
        ok: { type: boolean }
        challengeId: { type: string }
        accessToken: { type: string }
        refreshToken: { type: string }
        message: { type: string }
    AuthResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        message: { type: string }
    AuthError:
      type: object
      required: [ok, message]
      properties:
        ok: { type: boolean }
        message: { type: string }
    Citation:
      type: object
      required: [document, section, page, excerpt]
      properties:
        document: { type: string }
        section: { type: string }
        page: { type: number }
        excerpt: { type: string }
    RagAnswer:
      type: object
      required: [answer, confidence, citations]
      properties:
        answer: { type: string }
        confidence: { type: number }
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
    RagChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content: { type: string }
...
    RagSearchInput:
      type: object
      required: [query]
      properties:
        query: { type: string }
    RagChatInput:
      type: object
      required: [question]
      properties:
        question: { type: string }
        history:
          type: array
          items:
            $ref: "#/components/schemas/RagChatMessage"
    StorageSyncResult:
      type: object
      required: [scanned, indexed, updated, removed, errors]
      properties:
        scanned: { type: number }
        indexed: { type: number }
        updated: { type: number }
        removed: { type: number }
        errors:
          type: array
          items: { type: string }
...